pyproject.toml
[tool.poetry]
name = "email-summariser-agent"
version = "0.1.0"
description = "An agent that fetches and summarizes newsletters from Microsoft Outlook"
authors = ["yuewen_li@outlook.com"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
msal = "^1.25.0"
requests = "^2.31.0"
schedule = "^1.2.1"
python-dateutil = "^2.8.2"
ollama = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"